L2-992: Project Dropdown Accounts
