+++ frontend/src/lib/components/common/RouteModule.svelte
@@ -16,6 +16,7 @@
   const loadModule = async (): Promise<typeof SvelteComponent> => {
     switch (path) {
       case AppPath.Accounts:
+      case AppPath.LegacyAccounts:
         return (await import("../../../routes/Accounts.svelte")).default;
       case AppPath.LegacyNeurons:
       case AppPath.Neurons:
@@ -50,6 +51,9 @@
     },
     [AppPath.Accounts]: {
       title: $i18n.navigation.tokens,
+    },
+    [AppPath.LegacyAccounts]: {
+      title: $i18n.navigation.accounts,
     },
     [AppPath.LegacyNeurons]: {
       title: $i18n.navigation.neurons,
