+++ frontend/src/routes/ProjectDetail.svelte
@@ -30,7 +30,7 @@
 
   onMount(() => {
     if (!IS_TESTNET) {
-      routeStore.replace({ path: AppPath.Accounts });
+      routeStore.replace({ path: AppPath.LegacyAccounts });
     }
   });
